ai-guard-dog-backend/intelligence/requirements.txt
# AI Guard Dog Intelligence Layer - Python Dependencies

# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.26.0
requests>=2.31.0

# Google Gemini AI
google-generativeai>=0.3.2

# Data Processing & ML
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.26.3

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3

# Logging
loguru>=0.7.2
